labs/day3/requirements.txt
fastapi
uvicorn[standard]
pydantic
streamlit
sqlmodel
openai
python-dotenv